Plotting and update pump button

+++ datagraph.cpp
@@ -247,10 +247,10 @@ void DataGraph::removeAndCreateLineSeries()
 // #################### Public slots ###############
 void DataGraph::dataReadyForPlot(QStringList data)
 {
+    // qDebug() << data;
     for(int i = 0; i < recentData.size(); i++) {
         recentData[i] = (QPointF(static_cast<double>(data[0].toInt())/1000.0,
-                         static_cast<double>(data[i+1].toInt())));
-        //qDebug() << data[0].toInt() << data[i+1].toInt();
+                         static_cast<double>(data[i+1].toDouble())));
     }
 }
